meta-galileo: Add image-micro.bb
image-micro is smaller than clanton-full but more functional though
larger than clanton-tiny due to the addition of udev.

It's way too large for a real micro, which should be much smaller that
tiny, but this is just a starting point to begin paring things down.

The changes to grub.conf are some minor changes for development flow.
Basically, the first two menu items are for the virgin 0.8.0 kernels,
spi burned into flash, and clanton-full on the sdcard, but renamed to
live side-by-side with our development artifacts.

- bzImage is renamed to bzImage-full-clanton
- image-full-clanton.ext3 stays the same
- core-image-minimal-initramfs-clanton.cpio.gz is renamed to
  core-image-minimal-initramfs-image-full-clanton.cpio.gz, mainly for
  consistency.

The micro (development) artifacts are:
- bzImage
- image-micro-galileo.cpio.gz

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>