meta-galileo: Add mem= param to a more realistic value
We want to pretend that we have a small system, not the 256M that the
clanton board has.  The lowest I've been able to go so far was 22M
during boot though the actual run-time total is much less than that.

This was also the value needed during qemu testing...

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>