meta-galileo: Use the new xxx-disable kernel features
Use the new set of kernel features - most of these are really things
we want to be able to toggle at will and not necessarily include
unconditionally in the base.

root@galileo:/sys/kernel# cat virt_kmem
virtual kernel memory layout:
      .init : 0xc1181000 - 0xc11ab000   ( 168 kB)
      .data : 0xc112ae42 - 0xc1180640   ( 341 kB)
      .text : 0xc1000000 - 0xc112ae42   (1195 kB)

With this still have full TCP/IP networking but we lose mmcblk0 at
this point.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>