meta-galileo: Add some more xxx-disable kernel features
This basically represents the end of the low-hanging fruit i.e. what
we can toggle off without carving up the code.

Note that some of this isn't yet low-hanging - turning PROC off
results in hanging init, since a lot of that depends on
e.g. /proc/mounts, and it's still TBD which of net/inet we'll need.

Also added a toggle for kallsyms, since we want to be able to turn
that off and reboot with it on for making sense of OOPsen.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>