meta-galileo: Add packetif to lwip-echo
packetif is our lwip 'hardware' interface implementation.  'hardware'
in quotes because we're actually cheating and not interacting directly
with the ethernet interface, but building it on top of the kernel's
layer 2 implementation.  i.e.  We're only concerned with replacing the
kernel's TCP/UDP/IP layers with their userspace lwIP analogue and are
happy to have the kernel handle the rest.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>