meta-galileo: Misc refactor and new feature enablement
Refactor the kernel misc feature of note into separate features and
enable them.  This produces the following sizes:

  [trz@empanada linux-galileo-micro-build]$ size arch/x86/boot/vmlinux
     text          data                    bss        dec     hex filename
  1221702           192240                 2277376    3691318  385336     arch/x86/boot/vmlinux

  virtual kernel memory layout:
      .init : 0xc1133000 - 0xc115b000   ( 160 kB)
      .data : 0xc10f7060 - 0xc1132ac0   ( 238 kB)
      .text : 0xc1000000 - 0xc10f7060   ( 988 kB)

This is with LTO enabled, PTP_STMMAC disabled in galileo.cfg because
that completely breaks networking, everything disabled except
proc_sysctl, needed for dhcp, and proc, needed for a lot of things,
and rtnetlink and inet_raw also needed for bringing the interface up
and configuring via dhcp, etc.

Booted kernel is <1MB text with completely functional Linux networking
(modulo the net-diet config item, off of course in this
configuration).

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>